+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "feat-challenges-correction",
+  "name": "challenges-correction",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
@@ -8,8 +8,9 @@
   },
   "scripts": {
     "build": "babel src --extensions \".js,.ts\" --out-dir dist --copy-files",
-    "start": "node dist/index.js",
-    "dev": "ts-node-dev --inspect --transpile-only --ignore-watch node_module src/index.ts"
+    "dev": "npm-run-all -p dev:*",
+    "dev:server": "ts-node-dev --inspect --transpile-only --ignore-watch node_module src/index.ts",
+    "dev:queue": "ts-node-dev --inspect --transpile-only --ignore-watch node_module src/queue.ts"
   },
   "dependencies": {
     "axios": "^0.21.1",
@@ -38,6 +39,7 @@
     "@types/uuid": "^8.3.0",
     "babel-plugin-module-resolver": "^4.1.0",
     "babel-plugin-transform-typescript-metadata": "^0.3.2",
+    "npm-run-all": "^4.1.5",
     "ts-node-dev": "^1.1.6",
     "typescript": "^4.3.2"
   }
