+++ src/jobs/index.ts
@@ -1,0 +1,1 @@
+export { ChallengeCorrection } from './ChallengeCorrection';
