+++ src/queue.ts
@@ -1,0 +1,5 @@
+import 'dotenv/config';
+
+import { Queue } from './lib/Queue';
+
+Queue.process();
